+++ webserver/src/components/BasicSelect.js
@@ -4,7 +4,6 @@ import InputLabel from '@mui/material/InputLabel';
 import MenuItem from '@mui/material/MenuItem';
 import FormControl from '@mui/material/FormControl';
 import Select from '@mui/material/Select';
-import {SpacingWrapper} from "./Spacing.styles"
 
 function BasicSelect(arrayOfOptions, label) {
   const [state, setState] = React.useState('');
