+++ webserver/src/components/IdeaTable.js
@@ -1,4 +1,3 @@
-import react from 'react';
 import { DataGrid } from '@mui/x-data-grid';
 
 const rows= [
