--- webserver/src/components/MaterialButton.js
@@ -1,8 +1,0 @@
-import * as React from 'react';
-import Button from '@mui/material/Button';
-
-export default function MaterialButton() {
-  return (
-      <Button variant="contained">Contained</Button>
-  );
-}
