--- webserver/src/components/ReactHookFormSelect.js.js
@@ -1,32 +1,0 @@
-
-import { FormControl, InputLabel, Select } from "@mui/material";
-import { Controller } from "react-hook-form";
-
-
-// Test for Reusable FormSelect Component
-const ReactHookFormSelect = ({
-  name,
-  label,
-  control,
-  defaultValue,
-  children,
-  ...props
-}) => {
-  const labelId = `${name}-label`;
-  return (
-    <FormControl {...props}>
-      <InputLabel id={labelId}>{label}</InputLabel>
-      <Controller
-        as={
-          <Select labelId={labelId} label={label}>
-            {children}
-          </Select>
-        }
-        name={name}
-        control={control}
-        defaultValue={defaultValue}
-      />
-    </FormControl>
-  );
-};
-export default ReactHookFormSelect;
