--- webserver/src/components/Spacing.styles.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const SpacingWrapper = styled.aside`
-margin-bottom: 1%;
-padding: 1%;
-
-`;
